+++ src/client/gral/types.js
@@ -1,0 +1,3 @@
+// @flow
+
+export type KeyFilter = 'ALL' | 'UNUSED' | 'UNTRANSLATED' | 'FUZZY';